Add my solution please
Hi,
i have deployed my code on heroku, because i developed some views and some controller too. Feel free to not post the Heroku link if you think this is an "out of format" answer. Anyway the request to do in console work, so i think the basic solution is good.

bye,
Andrea.